+++ inc/session_utility.php
@@ -3347,7 +3347,9 @@ function texttodocount2($textid): string
         WHERE Ti2WordCount=1 AND Ti2WoID=0 AND Ti2TxID=$textid"
     );
     if ($c <= 0) {
-        return '<span title="To Do" class="status0">&nbsp;' . $c . '&nbsp;</span>'; 
+        return '<span title="No unknown word remaining" class="status0" 
+        style="padding: 0 5px; margin: 0 5px;">' . 
+        $c . '</span>'; 
     }
     $show_buttons = getSettingWithDefault('set-words-to-do-buttons');
     
@@ -3365,10 +3367,13 @@ function texttodocount2($textid): string
         $tl = $sl = "";
     }
     
-    $res = '<span title="To Do" class="status0">&nbsp;' . $c . '&nbsp;</span>&nbsp;' .
-    '<img src="icn/script-import.png" onclick="showRightFrames(\'bulk_translate_words.php?tid=' . 
-    $textid . '&offset=0&sl=' . $sl . '&tl=' . $tl . 
-    '\');" style="cursor: pointer;vertical-align:middle" title="Lookup New Words" alt="Lookup New Words" />&nbsp;&nbsp;&nbsp;';
+    $res = '<span title="Number of unknown words" class="status0" 
+    style="padding: 0 5px; margin: 0 5px;">' . $c . '</span>
+    <img src="icn/script-import.png" 
+    onclick="showRightFrames(\'bulk_translate_words.php?tid=' . $textid . 
+    '&offset=0&sl=' . $sl . '&tl=' . $tl . '\');" 
+    style="cursor: pointer; vertical-align:middle" title="Lookup New Words" 
+    alt="Lookup New Words" />';
     if ($show_buttons != 2) {
         $res .= '<input type="button" onclick="iknowall(' . $textid . ');" value=" Set All to Known " />'; 
     }